Module eventually_postgres::store

Contains an EventStore implementation using PostgreSQL as a backend data store.

Structs

EventStore

EventStore implementation using a PostgreSQL backend.

EventStoreBuilder

Builder type for EventStore instances.

EventStoreBuilderMigrated

Builder step for EventStore instances, after the database migration executed from EventStoreBuilder has been completed.

Enums

Error

Error type returned by the EventStore implementation, which is a newtype wrapper around tokio_postgres::Error.

Type Definitions

PoolResult

Result returning the connection pool Error type.

Result

Result returning the crate Error type.