Struct eventually_postgres::store::EventStore

pub struct EventStore<Id, Event, Tls> where
    Tls: MakeTlsConnect<Socket> + Clone + Send + Sync + 'static,
    <Tls as MakeTlsConnect<Socket>>::Stream: Send + Sync,
    <Tls as MakeTlsConnect<Socket>>::TlsConnect: Send,
    <<Tls as MakeTlsConnect<Socket>>::TlsConnect as TlsConnect<Socket>>::Future: Send,  { /* private fields */ }

EventStore implementation using a PostgreSQL backend.

This implementation uses tokio-postgres crate to interface with Postgres.

Check out EventStoreBuilder for examples to how initialize new instances of this type.

Considerations on the Id type

The Id type supplied to the EventStore has to be able to to_string() (so implement the std::fmt::Display trait) and to be parsed from a String (so to implement the std::convert::TryFrom<String> trait).

The error for the TryFrom<String> conversion has to be a std::error::Error, so as to be able to map it into an anyhow::Error generic error value.