Module eventually_postgres::subscription

Contains a persisted implementation of the Subscription trait using Postgres as the backend data source for its state.

Structs

Persistent
Subscription type with persistent state over a Postgres data source.

PersistentBuilder
Builder type for multiple Persistent Subscription instance.

Enums

Error
Error types returned by a Persistent Subscription.