Crate eventually_postgres

eventually type implementations for PostgreSQL.

Event Store

This crate includes an EventStore implementation using PostgreSQL as backend data source.

Example usage:

// Open a connection with bb8.
let pg_manager = bb8_postgres::PostgresConnectionManager::new_from_stringlike(
    "postgres://user@pass:localhost:5432/db",
    tokio_postgres::NoTls,
)
.map_err(|err| {
    eprintln!("Failed configuring a Postgres Connection pool: {}", err);
    err
})?;
let pool = bb8::Pool::builder().build(pg_manager).await?;

// A domain event example -- it is deliberately simple.
#[derive(Debug, Clone)]
struct SomeEvent;

// Use an EventStoreBuilder to build multiple EventStore instances.
let builder = EventStoreBuilder::migrate_database(pool.clone())
    .await?
    .builder(pool);

// Event store for the events.
//
// When building an new EventStore instance, a type name is always needed
// to distinguish between different aggregates.
//
// You can also use std::any::type_name for that.
let store = builder.build::<String, SomeEvent>("aggregate-name").await?;

Re-exports

pub use store::*;

pub use subscriber::*;

pub use subscription::*;

Modules

store

Contains an EventStore implementation using PostgreSQL as a backend data store.

subscriber

Contains an EventSubscriber implementation using PostgreSQL as a backend data store and NOTIFY/LISTEN functionality to power the EventStream.

subscription

Contains a persisted implementation of the Subscription trait using Postgres as the backend data source for its state.