Struct eventually_postgres::EventStore

pub struct EventStore<Id, Event> { /* fields omitted */ }

EventStore implementation using a PostgreSQL backend.

This implementation uses tokio-postgres crate to interface with Postgres.

Check out EventStoreBuilder for examples to how initialize new instances of this type.

Methods

impl<Id, Event> EventStore<Id, Event> where
Id: ToString + Eq + Send + Sync,

pub async fn create_stream<'_>(&'_ self) -> Result<(), Error>

Creates a new table in the database for the provided Stream name during initialization.

Check out EventStoreBuilder for more information.